pyproject.toml
[project]
name = "krddevbot"
version = "0.1.0"
description = ""
authors = [
    {name = "TyVik", email = "tyvik@krd.dev"}
]
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "python-telegram-bot[job-queue]>=20.8",
    "sentry-sdk>=1.39.2",
]

[project.urls]
Homepage = "https://t.me/krddevdays_chat"
Repository = "https://github.com/krddevdays/bot"

[project.scripts]
pdm = "pdm.core:main"

[tool.pdm.scripts]
_.env_file = ".env"
bot = "python -m krddevbot"
tests = "pytest tests"

# dev run: nodemon -e py --exec pdm run bot

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.0.0",
    "flake8>=5.0.0", 
]